+++ lib/src/model/message_context.dart
@@ -5,20 +5,28 @@ import 'package:televerse/telegram.dart';
 class MessageContext {
   final int userId;
   final int chatId;
-  final String? text;
-  final int? editMessageId;
   final String? username;
+  final String? locale;
+  final String? text;
+
+  /// Id of the message to which user is replying to
+  /// Common usecase is to update that message by removing buttons
+  final int? editMessageId;
   final StickerModel? sticker; //todo should remove and use mediaFiles instead?
   final List<MediaFile>? mediaFiles;
   final PreCheckoutQuery? preCheckoutInfo; //todo should not use model from televerse?
   final SuccessfulPayment? successfulPayment;
 
+  final Message? original;
+
   MessageContext({
     required this.userId,
     required this.chatId,
+    required this.original,
+    this.username,
+    this.locale,
     this.text,
     this.editMessageId,
-    this.username,
     this.sticker,
     this.mediaFiles,
     this.preCheckoutInfo,